backend/requirements.txt
fastapi
uvicorn
debugpy
google-genai>=0.3.0
python-dotenv>=1.0.1